redefinition of hooks, context, and react components
